# [Improvement] Update `__cache` of ancestor keys