@ianstormtaylor 2018-02-07
@zhujinxuan it's weird that `fromJSON` is seeing that change. But it also seems like this isn't really speeding things up much?