@zhujinxuan 2018-02-07
@ianstormtaylor I think one reason is that the benchmark is using the very simple document for test.  There is not many tables or quotes or inlines in these benchmarks. 

I am thinking about how to use `tests` document as benchmark test.