@ianstormtaylor 2018-02-09
@zhujinxuan sounds good. I'd need to see proof that this has pretty decent performance boosts, because it's a very complex thing to change and keep track of mentally, so I'd rather not make this change if possible. If it happens to have huge performance gains I'll consider it though.