@zhujinxuan 2018-03-20
Close this issue because this function is rewrited with better interface in #1713